+++ events_specialize.go
@@ -176,6 +176,8 @@ func (pev *Event) Specialize() DetailedEvent {
 		e = &EventLinkGithubType{Event: ev}
 	case EventUnlinkGithub:
 		e = &EventUnlinkGithubType{Event: ev}
+	case EventDeleteProject:
+		e = &EventDeleteProjectType{Event: ev}
 	case EventNewProject:
 		e = &EventNewProjectType{Event: ev}
 	default: