+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## To Be Released
 
+* feat(projects): add delete project event
 * feat(projects): add new project created event
 
 ## 8.1.0